+++ lib/rspec/mocks/method_signature_verifier.rb
@@ -1,5 +1,3 @@
-RSpec::Support.require_rspec_mocks 'ruby_features'
-
 module RSpec
   module Mocks
     # Extracts info about the number of arguments and allowed/required