+++ Changelog.md
@@ -23,6 +23,8 @@ Bug Fixes:
   not setting an expectation for the last message in the chain.
   (Jonathan del Strother)
 * Allow verifying partial doubles to have private methods stubbed. (Xavier Shay)
+* Fix bug with allowing/expecting messages on Class objects which have had
+  their singleton class prepended to. (Jon Rowe)
 
 ### 3.0.0.beta2 / 2014-02-17
 [Full Changelog](http://github.com/rspec/rspec-mocks/compare/v3.0.0.beta1...v3.0.0.beta2)